Resize images (better quality)
Re-obtained the images at the largest possible dimensions and then resized
them as necessary. Might compress them later, but compressing then resizing
seems to have made the images of lower quality than is optimal. Or it might
have something to do with the quality of the original images. In any case,
it's better now.

This commit also changes the printer statuses icon on the homepage.